ProjectCards correctly displaying hard coded card
