+++ src/Styles/ProjectCards.scss
@@ -3,6 +3,8 @@
 }
 :root {
     --fontBlack: #6E6E6E;
+    --buttonColor: #91B0B8;
+    --buttonColorHover: #66878f;
 }
 
 .project-card-container {
@@ -52,11 +54,15 @@
         width: 80px; 
         height: 30px;
         color: white;
-        background-color: #91B0B8;
+        background-color: var(--buttonColor);
         border-radius: 4px;
         border: none;
         margin: 15px 10px 0 10px;
         font-size: 1.1em;
+    }
+    .project-card-buttons :hover {
+        cursor: pointer;
+        background-color: var(--buttonColorHover);
     }
 
 }
